fix: remove empty nodeModules option from bundling
Including this option, even as an empty list, will lead to the NPM lock file
being copied over to the asset's directory during synthesis, thus changing
the asset hash every time the lock file is updated. This is not our
desired behavior, at least not when not relying on any external
packages.

The behavior is triggered by:
https://github.com/aws/aws-cdk/blob/9d61cb97e0f52460020ea6f3e780070d98503b19/packages/aws-cdk-lib/aws-lambda-nodejs/lib/bundling.ts#L225